requirements.txt
matplotlib==3.9.2
nicegui==2.5.0
numpy==2.1.2
Pillow==11.0.0
tqdm==4.66.5
